Cleanup / consolidation
Only meaningful code changes are deleting the extension methods on OpenAIClient (keeping the ones on ChatClient / EmbeddingClient), and renaming the AsChatClient / AsEmbeddingGenerator extension methods to be AsIChatClient / AsIEmbeddingGenerator. Otherwise, "ChatClient" is confusing with the OpenAI type of the same name.
