+++ src/Libraries/Microsoft.Extensions.AI.Abstractions/Contents/DataContent.cs
@@ -110,7 +110,7 @@ public class DataContent : AIContent
     /// </summary>
     /// <param name="data">The byte contents.</param>
     /// <param name="mediaType">The media type (also known as MIME type) represented by the content.</param>
-    /// <exception cref="ArgumentNullException"><paramref name="mediaType"/> is null.</exception>
+    /// <exception cref="ArgumentNullException"><paramref name="mediaType"/> is <see langword="null"/>.</exception>
     /// <exception cref="ArgumentException"><paramref name="mediaType"/> is empty or composed entirely of whitespace.</exception>
     public DataContent(ReadOnlyMemory<byte> data, string mediaType)
     {
