+++ src/Libraries/Microsoft.Extensions.AI.AzureAIInference/AzureAIInferenceEmbeddingGenerator.cs
@@ -36,31 +36,31 @@ internal sealed class AzureAIInferenceEmbeddingGenerator :
 
     /// <summary>Initializes a new instance of the <see cref="AzureAIInferenceEmbeddingGenerator"/> class.</summary>
     /// <param name="embeddingsClient">The underlying client.</param>
-    /// <param name="modelId">
+    /// <param name="defaultModelId">
     /// The ID of the model to use. This can also be overridden per request via <see cref="EmbeddingGenerationOptions.ModelId"/>.
     /// Either this parameter or <see cref="EmbeddingGenerationOptions.ModelId"/> must provide a valid model ID.
     /// </param>
-    /// <param name="dimensions">The number of dimensions to generate in each embedding.</param>
+    /// <param name="defaultModelDimensions">The number of dimensions to generate in each embedding.</param>
     /// <exception cref="ArgumentNullException"><paramref name="embeddingsClient"/> is <see langword="null"/>.</exception>
-    /// <exception cref="ArgumentException"><paramref name="modelId"/> is empty or composed entirely of whitespace.</exception>
-    /// <exception cref="ArgumentOutOfRangeException"><paramref name="dimensions"/> is not positive.</exception>
+    /// <exception cref="ArgumentException"><paramref name="defaultModelId"/> is empty or composed entirely of whitespace.</exception>
+    /// <exception cref="ArgumentOutOfRangeException"><paramref name="defaultModelDimensions"/> is not positive.</exception>
     public AzureAIInferenceEmbeddingGenerator(
-        EmbeddingsClient embeddingsClient, string? modelId = null, int? dimensions = null)
+        EmbeddingsClient embeddingsClient, string? defaultModelId = null, int? defaultModelDimensions = null)
     {
         _ = Throw.IfNull(embeddingsClient);
 
-        if (modelId is not null)
+        if (defaultModelId is not null)
         {
-            _ = Throw.IfNullOrWhitespace(modelId);
+            _ = Throw.IfNullOrWhitespace(defaultModelId);
         }
 
-        if (dimensions is < 1)
+        if (defaultModelDimensions is < 1)
         {
-            Throw.ArgumentOutOfRangeException(nameof(dimensions), "Value must be greater than 0.");
+            Throw.ArgumentOutOfRangeException(nameof(defaultModelDimensions), "Value must be greater than 0.");
         }
 
         _embeddingsClient = embeddingsClient;
-        _dimensions = dimensions;
+        _dimensions = defaultModelDimensions;
 
         // https://github.com/Azure/azure-sdk-for-net/issues/46278
         // The endpoint isn't currently exposed, so use reflection to get at it, temporarily. Once packages
@@ -69,7 +69,7 @@ internal sealed class AzureAIInferenceEmbeddingGenerator :
         var providerUrl = typeof(EmbeddingsClient).GetField("_endpoint", BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
             ?.GetValue(embeddingsClient) as Uri;
 
-        _metadata = new EmbeddingGeneratorMetadata("az.ai.inference", providerUrl, modelId, dimensions);
+        _metadata = new EmbeddingGeneratorMetadata("az.ai.inference", providerUrl, defaultModelId, defaultModelDimensions);
     }
 
     /// <inheritdoc />
