--- src/Libraries/Microsoft.Extensions.AI.OpenAI/OpenAIModelMapper.ChatMessage.cs
@@ -1,129 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System;
-using System.Collections.Generic;
-using System.Text.Json;
-using OpenAI.Chat;
-
-namespace Microsoft.Extensions.AI;
-
-internal static partial class OpenAIModelMappers
-{
-    public static ChatRole ChatRoleDeveloper { get; } = new ChatRole("developer");
-
-    /// <summary>Converts an Extensions chat message enumerable to an OpenAI chat message enumerable.</summary>
-    public static IEnumerable<OpenAI.Chat.ChatMessage> ToOpenAIChatMessages(IEnumerable<ChatMessage> inputs, JsonSerializerOptions options)
-    {
-        // Maps all of the M.E.AI types to the corresponding OpenAI types.
-        // Unrecognized or non-processable content is ignored.
-
-        foreach (ChatMessage input in inputs)
-        {
-            if (input.Role == ChatRole.System ||
-                input.Role == ChatRole.User ||
-                input.Role == ChatRoleDeveloper)
-            {
-                var parts = ToOpenAIChatContent(input.Contents);
-                yield return
-                    input.Role == ChatRole.System ? new SystemChatMessage(parts) { ParticipantName = input.AuthorName } :
-                    input.Role == OpenAIModelMappers.ChatRoleDeveloper ? new DeveloperChatMessage(parts) { ParticipantName = input.AuthorName } :
-                    new UserChatMessage(parts) { ParticipantName = input.AuthorName };
-            }
-            else if (input.Role == ChatRole.Tool)
-            {
-                foreach (AIContent item in input.Contents)
-                {
-                    if (item is FunctionResultContent resultContent)
-                    {
-                        string? result = resultContent.Result as string;
-                        if (result is null && resultContent.Result is not null)
-                        {
-                            try
-                            {
-                                result = JsonSerializer.Serialize(resultContent.Result, options.GetTypeInfo(typeof(object)));
-                            }
-                            catch (NotSupportedException)
-                            {
-                                // If the type can't be serialized, skip it.
-                            }
-                        }
-
-                        yield return new ToolChatMessage(resultContent.CallId, result ?? string.Empty);
-                    }
-                }
-            }
-            else if (input.Role == ChatRole.Assistant)
-            {
-                AssistantChatMessage message = new(ToOpenAIChatContent(input.Contents))
-                {
-                    ParticipantName = input.AuthorName
-                };
-
-                foreach (var content in input.Contents)
-                {
-                    if (content is FunctionCallContent callRequest)
-                    {
-                        message.ToolCalls.Add(
-                            ChatToolCall.CreateFunctionToolCall(
-                                callRequest.CallId,
-                                callRequest.Name,
-                                new(JsonSerializer.SerializeToUtf8Bytes(
-                                    callRequest.Arguments,
-                                    options.GetTypeInfo(typeof(IDictionary<string, object?>))))));
-                    }
-                }
-
-                if (input.AdditionalProperties?.TryGetValue(nameof(message.Refusal), out string? refusal) is true)
-                {
-                    message.Refusal = refusal;
-                }
-
-                yield return message;
-            }
-        }
-    }
-
-    /// <summary>Converts a list of <see cref="AIContent"/> to a list of <see cref="ChatMessageContentPart"/>.</summary>
-    private static List<ChatMessageContentPart> ToOpenAIChatContent(IList<AIContent> contents)
-    {
-        List<ChatMessageContentPart> parts = [];
-        foreach (var content in contents)
-        {
-            switch (content)
-            {
-                case TextContent textContent:
-                    parts.Add(ChatMessageContentPart.CreateTextPart(textContent.Text));
-                    break;
-
-                case UriContent uriContent when uriContent.HasTopLevelMediaType("image"):
-                    parts.Add(ChatMessageContentPart.CreateImagePart(uriContent.Uri));
-                    break;
-
-                case DataContent dataContent when dataContent.HasTopLevelMediaType("image"):
-                    parts.Add(ChatMessageContentPart.CreateImagePart(BinaryData.FromBytes(dataContent.Data), dataContent.MediaType));
-                    break;
-
-                case DataContent dataContent when dataContent.HasTopLevelMediaType("audio"):
-                    var audioData = BinaryData.FromBytes(dataContent.Data);
-                    if (dataContent.MediaType.Equals("audio/mpeg", StringComparison.OrdinalIgnoreCase))
-                    {
-                        parts.Add(ChatMessageContentPart.CreateInputAudioPart(audioData, ChatInputAudioFormat.Mp3));
-                    }
-                    else if (dataContent.MediaType.Equals("audio/wav", StringComparison.OrdinalIgnoreCase))
-                    {
-                        parts.Add(ChatMessageContentPart.CreateInputAudioPart(audioData, ChatInputAudioFormat.Wav));
-                    }
-
-                    break;
-            }
-        }
-
-        if (parts.Count == 0)
-        {
-            parts.Add(ChatMessageContentPart.CreateTextPart(string.Empty));
-        }
-
-        return parts;
-    }
-}
