+++ test/Libraries/Microsoft.Extensions.AI.AzureAIInference.Tests/IntegrationTestHelpers.cs
@@ -21,13 +21,13 @@ internal static class IntegrationTestHelpers
         TestRunnerConfiguration.Instance["AzureAIInference:Endpoint"] ??
         "https://api.openai.com/v1";
 
-    /// <summary>Gets a <see cref="ChatResponsesClient"/> to use for testing, or null if the associated tests should be disabled.</summary>
+    /// <summary>Gets a <see cref="ChatResponsesClient"/> to use for testing, or <see langword="null"/> if the associated tests should be disabled.</summary>
     public static ChatCompletionsClient? GetChatCompletionsClient() =>
         _apiKey is string apiKey ?
             new ChatCompletionsClient(new Uri(_endpoint), new AzureKeyCredential(apiKey), CreateOptions()) :
             null;
 
-    /// <summary>Gets an <see cref="EmbeddingsClient"/> to use for testing, or null if the associated tests should be disabled.</summary>
+    /// <summary>Gets an <see cref="EmbeddingsClient"/> to use for testing, or <see langword="null"/> if the associated tests should be disabled.</summary>
     public static EmbeddingsClient? GetEmbeddingsClient() =>
         _apiKey is string apiKey ?
             new EmbeddingsClient(new Uri(_endpoint), new AzureKeyCredential(apiKey), CreateOptions()) :
