+++ test/Libraries/Microsoft.Extensions.AI.Evaluation.Integration.Tests/Setup.cs
@@ -17,13 +17,13 @@ internal static class Setup
     internal static ChatConfiguration CreateChatConfiguration()
     {
         var endpoint = new Uri(Settings.Current.Endpoint);
-        AzureOpenAIClientOptions options = new AzureOpenAIClientOptions();
+        AzureOpenAIClientOptions options = new();
         AzureOpenAIClient azureClient =
             OfflineOnly
                 ? new AzureOpenAIClient(endpoint, new ApiKeyCredential("Bogus"), options)
                 : new AzureOpenAIClient(endpoint, new DefaultAzureCredential(), options);
 
-        IChatClient chatClient = azureClient.AsChatClient(Settings.Current.DeploymentName);
+        IChatClient chatClient = azureClient.GetChatClient(Settings.Current.DeploymentName).AsIChatClient();
         Tokenizer tokenizer = TiktokenTokenizer.CreateForModel(Settings.Current.ModelName);
         IEvaluationTokenCounter tokenCounter = tokenizer.ToTokenCounter(inputTokenLimit: 6000);
         return new ChatConfiguration(chatClient, tokenCounter);
